classifier/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
scikit-learn==1.7.1
joblib==1.5.2
numpy==2.3.3
